+++ lib/rspec/mocks/space.rb
@@ -58,9 +58,9 @@ module RSpec
                         when TestDouble then object.__build_mock_proxy(expectation_ordering)
                         else
                           if RSpec::Mocks.configuration.verify_partial_doubles?
-                            VerifyingPartialMockProxy.new(object, expectation_ordering)
+                            VerifyingPartialDoubleProxy.new(object, expectation_ordering)
                           else
-                            PartialMockProxy.new(object, expectation_ordering)
+                            PartialDoubleProxy.new(object, expectation_ordering)
                           end
                         end
         end